+++ apps/server/src/modules/account/repo/account.repo.ts
@@ -6,8 +6,7 @@ import { Counted, EntityId } from '@shared/domain/types';
 import { BaseDomainObjectRepo } from '@shared/repo/base-domain-object.repo';
 import { Account, AccountRepo } from '../domain';
 import { AccountEntity } from './account.entity';
-import { AccountEntityToDoMapper } from './mapper';
-import { AccountDoToEntityMapper } from './mapper/account-do-to-entity.mapper';
+import { AccountDoToEntityMapper, AccountEntityToDoMapper } from './mapper';
 import { AccountScope } from './scope/account-scope';
 
 @Injectable()
