+++ apps/server/src/modules/account/repo/mapper/account-entity-to-do.mapper.ts
@@ -1,5 +1,5 @@
 import { Counted } from '@shared/domain/types';
-import { Account } from '../../domain/do/account';
+import { Account } from '../../domain';
 import { AccountEntity } from '../account.entity';
 
 export class AccountEntityToDoMapper {
